% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adj2edges}
\alias{adj2edges}
\title{Turn a adjacency matrix into a matrix of edges}
\usage{
adj2edges(adj)
}
\arguments{
\item{adj}{adjacency matrix}
}
\value{
m x 2 matix where each row r denotes an edge from r[,1] to r[,2] and m is the number of edges in adj
}
\description{
Turn a adjacency matrix into a matrix of edges
}
\note{
equivalent to which(adj, arr.ind = T)
}
