% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{allOrderPermutations}
\alias{allOrderPermutations}
\title{Compute all causal orderings of a number of nodes}
\usage{
allOrderPermutations(p)
}
\arguments{
\item{p}{Number of nodes that are to be permutated.}
}
\value{
List of order vectors (\eqn{pi})
}
\description{
Compute all causal orderings of a number of nodes
}
