% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{areAllCausalOrdersCompatible}
\alias{areAllCausalOrdersCompatible}
\title{Check if all causal orderings of a DAG \code{G} compatible with DAG \code{H}
\eqn{\pi \in \Pi(H) \forall \pi \in \Pi(G)}}
\usage{
areAllCausalOrdersCompatible(G, H)
}
\arguments{
\item{G, }{H Adjacenzy matrices of the two DAGs.}
}
\value{
TRUE if there is causal a ordering compatible with both DAGs
}
\description{
Check if all causal orderings of a DAG \code{G} compatible with DAG \code{H}
\eqn{\pi \in \Pi(H) \forall \pi \in \Pi(G)}
}
