% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.cam.one_sided}
\alias{bootstrap.cam.one_sided}
\title{Test for the existence of a total causal effect from node i to node j
H0: there is no causal effekt from i to j}
\usage{
bootstrap.cam.one_sided(
  X,
  ij,
  B = 100,
  nodeModelName = NULL,
  nodeModelPars = NULL,
  verbose = 0,
  lambda = 1,
  bs_lvl0 = FALSE,
  mode = c("H0: i -/->j", "H0: i-->j")
)
}
\arguments{
\item{X}{data used for}

\item{lambda}{use 0.5 for mid pvalue  see 1. Franck, W. E. P-values For Discrete Test Statistics. Biometrical J. 28, 403–406 (1986).}
}
\description{
Test for the existence of a total causal effect from node i to node j
H0: there is no causal effekt from i to j
}
