% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.cam.one_sided_direct}
\alias{bootstrap.cam.one_sided_direct}
\title{H0: i -/->j}
\usage{
bootstrap.cam.one_sided_direct(
  X,
  B = 99,
  nodeModelName = NULL,
  nodeModelPars = NULL,
  verbose = 0,
  lambda = 1,
  seed_ = NULL
)
}
\description{
H0: i -/->j
}
