% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{causalOrderToAdjacency}
\alias{causalOrderToAdjacency}
\title{Calculate adjacency matrix of the fully connected DAG corresponiding to a causal order}
\usage{
causalOrderToAdjacency(causalOrder)
}
\arguments{
\item{causalOrder}{Vector of length \eqn{p} where the \eqn{i}'th element is \eqn{\pi(i)}.}
}
\description{
Calculate adjacency matrix of the fully connected DAG corresponiding to a causal order
}
