% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkCausalOrder}
\alias{checkCausalOrder}
\title{Check if the causal ordering of an DAG (estDAG) matches the ordering of an other DAG (trueDAG)
using their adjacency matrices}
\usage{
checkCausalOrder(estDAG, trueDAG)
}
\arguments{
\item{estDAG}{Adjacenzy matrix of the DAG whichs order is to be checked}

\item{trueDAG}{Adjacenzy matrix of the DAG which defines the set of "correct" orderings}
}
\value{
TRUE if the causal ordering matches the constrains, otherwise FALSE
}
\description{
Check if the causal ordering of an DAG (estDAG) matches the ordering of an other DAG (trueDAG)
using their adjacency matrices
}

