% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incEdge.R
\name{computeScoreMatEntry}
\alias{computeScoreMatEntry}
\title{Computes an entry of the score matrix that is the gain in scoreFunctions value upon addition of an edge from i to j}
\usage{
computeScoreMatEntry(
  existingParOfJ,
  nodeModelName,
  nodeModelPars,
  scoreFunction,
  X,
  verbose,
  i,
  j,
  oldScore
)
}
\description{
Computes an entry of the score matrix that is the gain in scoreFunctions value upon addition of an edge from i to j
}
\note{
This is an auxiliary file for the function CAM in package CAM.
}
\references{
P. B\"uhlmann, J. Peters, J. Ernest: CAM: Causal Additive
Models, high-dimensional Order Search and Penalized Regression Annals of
Statistics 42:2526-2556, 2014.
}
\seealso{
\code{\link[CAM]{CAM}}
}
\author{
Jonas Peters <jonas.peters@tuebingen.mpg.de> and Jan Ernest
<ernest@stat.math.ethz.ch>
}
