% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeScoreMatParallel.R
\name{computeScoreMatParallel}
\alias{computeScoreMatParallel}
\title{auxiliary file for CAM: Computes the initial score matrix.}
\usage{
computeScoreMatParallel(rowParents, scoreName, X, selMat, output, node2, i,
  parsScore, intervMat, intervData)
}
\description{
This is an auxiliary file for the function CAM in package CAM.
}
\details{
This is an auxiliary file for the function CAM in package CAM.
}
\note{
This is an auxiliary file for the function CAM in package CAM.
}
\author{
Jonas Peters <jonas.peters@tuebingen.mpg.de> and Jan Ernest
<ernest@stat.math.ethz.ch>
}
\references{
P. B\"uhlmann, J. Peters, J. Ernest: CAM: Causal Additive
Models, high-dimensional Order Search and Penalized Regression Annals of
Statistics 42:2526-2556, 2014.
}
\seealso{
\code{\link[CAM]{CAM}}
}

