% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{d.3d_featureplot}
\alias{d.3d_featureplot}
\title{3d plot baaaaby}
\usage{
d.3d_featureplot(
  data,
  feature,
  reduction,
  type,
  ident,
  size,
  max.cutoff,
  point
)
}
\description{
3d plot baaaaby
}
