% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dagToCausalOrder}
\alias{dagToCausalOrder}
\title{Calculate a causal order of a DAG \code{causalDAG}}
\usage{
dagToCausalOrder(causalDAG)
}
\arguments{
\item{causalDAG}{Adjacency matrix of the DAG whose order is to be computed.}
}
\value{
\eqn{\pi}(\code{causalDAG})
}
\description{
Calculate a causal order of a DAG \code{causalDAG}
}
