% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{existsCompatibleCausalOrder}
\alias{existsCompatibleCausalOrder}
\title{Check if there is causal ordering compatible with both of two DAGs
\eqn{\exists \pi \in \Pi(G) \cup \Pi(H)}}
\usage{
existsCompatibleCausalOrder(G, H)
}
\arguments{
\item{G, }{H Adjacenzy matrices of the two DAGs.}
}
\value{
TRUE if there is causal a ordering compatible with both DAGs
}
\description{
Check if there is causal ordering compatible with both of two DAGs
\eqn{\exists \pi \in \Pi(G) \cup \Pi(H)}
}
