% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getCompatibleCausalOrder}
\alias{getCompatibleCausalOrder}
\title{Get compatible causal order}
\usage{
getCompatibleCausalOrder(estDAG, trueDAG)
}
\arguments{
\item{estDAG}{Adjacenzy matrix of the DAG whichs order is to be checked.}

\item{trueDAG}{Adjacenzy matrix of the DAG which defines the set of "correct" orderings.}
}
\value{
TRUE if the causal ordering matches the constrains, otherwise FALSE
}
\description{
Get compatible causal order
}
