% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPathMatrix}
\alias{getPathMatrix}
\title{Compute the path matrix of a graph G from the adjacency matrix of G}
\usage{
getPathMatrix(adj)
}
\arguments{
\item{adj}{adjacency matrix to compute the pathmatrix for. Must be a \eqn{p\times{}p} matrix that
is TRUE at the edges of G and FALSE else where}
}
\description{
Compute the path matrix of a graph G from the adjacency matrix of G
}

