% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAM.R
\name{preliminaryNeighborSel}
\alias{preliminaryNeighborSel}
\title{A matrix selMat is constructed. Entry (i,j) being one means that i is a possible parent of j.}
\usage{
preliminaryNeighborSel(
  X,
  method = NULL,
  modelPars = NULL,
  selectionPars = NULL,
  intervData = FALSE,
  intervMat = NULL,
  numCores = 1,
  verbose = F
)
}
\description{
A matrix selMat is constructed. Entry (i,j) being one means that i is a possible parent of j.
}
