% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selGam.R
\name{selGam}
\alias{selGam}
\title{auxiliary file for CAM: selection based on gam (mgcv)}
\usage{
selGam(X, pars = list(cutOffPVal = 0.001, numBasisFcts = 10),
  output = FALSE, k)
}
\description{
auxiliary file for CAM: selection based on gam (mgcv)
}

