% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selGamBoost.R
\name{selGamBoost}
\alias{selGamBoost}
\title{auxiliary file for CAM: selection based on gam boost (mboost)}
\usage{
selGamBoost(X, pars = list(atLeastThatMuchSelected = 0.02,
  atMostThatManyNeighbors = 10), output = FALSE, k)
}
\description{
auxiliary file for CAM: selection based on gam boost (mboost)
}

