% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selLasso.R
\name{selLasso}
\alias{selLasso}
\title{auxiliary file for CAM: selection based on lasso (glmnet)}
\usage{
selLasso(X, pars = list(), output = FALSE, k)
}
\description{
auxiliary file for CAM: selection based on lasso (glmnet)
}

