% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selLm.R
\name{selLm}
\alias{selLm}
\title{auxiliary file for CAM: selection based on lm}
\usage{
selLm(X, pars = list(cutOffPVal = 0.001), output = FALSE, k)
}
\description{
auxiliary file for CAM: selection based on lm
}

