% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selLmBoost.R
\name{selLmBoost}
\alias{selLmBoost}
\title{auxiliary file for CAM: selection based on lm (mboost)}
\usage{
selLmBoost(X, pars = list(atLeastThatMuchSelected = 0.02,
  atMostThatManyNeighbors = 10), output = FALSE, k)
}
\description{
auxiliary file for CAM: selection based on lm (mboost)
}

