% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{simulate_bootCAM_null_resampling}
\alias{simulate_bootCAM_null_resampling}
\title{H0: i -/->j}
\usage{
simulate_bootCAM_null_resampling(
  X,
  B = 99,
  parsScore = list(),
  nodeModelPars = NULL,
  verbose = FALSE,
  seed_ = NULL
)
}
\description{
H0: i -/->j
}
