% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/incEdge.R
\name{updateScoreMat}
\alias{updateScoreMat}
\title{auxiliary file for CAM: Updates the score matrix after having added edge i
-> j to the graph.}
\usage{
updateScoreMat(scoreMat, X, nodeModelName, nodeModelPars, scoreFunction, i, j,
  scoreNodes, Adj, maxNumParents, intervMat, intervData, verbose = FALSE,
  numCores = 0)
}
\arguments{
\item{scoreMat}{the current score matrix that has to be updated}

\item{X}{nxp matrix of training inputs (n data points, p dimensions)}

\item{nodeModelName}{vector containing current scores of each node}

\item{nodeModelPars}{additional parameters can be supported to the score
function.}

\item{i}{starting point of the edge i->j that has been added}

\item{j}{end point of the edge i->j that has been added}

\item{Adj}{adjacency matrix of the graph}

\item{maxNumParents}{specifies the maximal number of parents that are
allowed in the model.}

\item{intervMat}{the matrix intervMat has the same dimension as X. entry
(i,j) == TRUE indicates that in experiment i, variable j has been intervened
on.}

\item{intervData}{boolean that indicates whether we use interventional data.}

\item{verbose}{boolean indicating whether information about the progress is
written to the console.}

\item{numCores}{specifies the number of cores that can be used for
computation.}

\item{scoreName}{specifies the model type which is used to compute the
score. Default is "SEMGAM" which assumes a generalized additive model class.
Other options include "SEMLIN" which fits a linear model.}
}
\value{
\item{scoreMat}{the updated score matrix.}
}
\description{
auxiliary file for CAM: Updates the score matrix after having added edge i
-> j to the graph.
}
\note{
This is an auxiliary file for CAM.
}
\author{
J. Peters (jonas.peters@tuebingen.mpg.de) and J. Ernest
(ernest@stat.math.ethz.ch)
}
\references{
P. B\"uhlmann, J. Peters, J. Ernest: CAM: Causal Additive
Models, high-dimensional Order Search and Penalized Regression Annals of
Statistics 42:2526-2556, 2014.
}
\seealso{
\code{\link[CAM]{CAM}}
}

