\name{CAM-package}
\alias{CAM-package}
\docType{package}
\title{
Causal Additive Model (package)
}
\description{
fits a causal additive model using the CAM algorithm, see references below
}
\details{
\tabular{ll}{
Package: \tab CAM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-03\cr
License: \tab GPL\cr
}
Fitting of a causal additive model. The algorithm can be called with \code{CAM}.
}
\author{
Jonas Peters and Jan Ernest
}
\references{
P. Bühlmann, J. Peters, J. Ernest: "CAM: Causal Additive Models, high-dimensional 
order search and penalized regression", Annals of Statistics 42:2526-2556, 2014.
}

