\name{computeScoreMatParallel}
\alias{computeScoreMatParallel}
\title{
auxiliary file for CAM: Computes the initial score matrix.
}
\description{
This is an auxiliary file for the function CAM in package CAM.
}
\details{
This is an auxiliary file for the function CAM in package CAM.
}
\references{
P. B\"uhlmann, J. Peters, J. Ernest: 
CAM: Causal Additive Models, high-dimensional Order Search and Penalized Regression
Annals of Statistics 42:2526-2556, 2014.
}
\author{
Jonas Peters <jonas.peters@tuebingen.mpg.de> and Jan Ernest <ernest@stat.math.ethz.ch>
}
\note{
This is an auxiliary file for the function CAM in package CAM.
}
\seealso{
\code{\link[CAM]{CAM}}
}




